Split new app config into smaller files
