+++ internal/app/serde_test.go
